Add script for Cloudflare Beacon analytics
The Cloudflare Beacon analytics script has been added to the app's layout file. This was added to get a clearer understanding of the website’s performance as well as gain better insights of user interactions within the application. Now, once the page is loaded, it logs 'Cloudflare Beacon has loaded'.